+++ packages/create/src/generators/app-lit-element-ts/templates/static-testing/web-test-runner.config.mjs
@@ -1,0 +1,4 @@
+export default {
+  files: 'out-tsc/test/**/*.test.js',
+  nodeResolve: true
+};