+++ packages/create/src/generators/testing-wtr-ts/templates/static/web-test-runner.config.mjs
@@ -1,0 +1,4 @@
+export default {
+  files: 'dist/**/test/**/*.test.js',
+  nodeResolve: true
+};