+++ packages/semantic-dom-diff/get-diffable-html.js
@@ -76,6 +76,9 @@ export function getDiffableHTML(html, options = {}) {
   const ignoreTags = [...(options.ignoreTags || []), ...DEFAULT_IGNORE_TAGS];
   const ignoreChildren = options.ignoreChildren || [];
   const stripEmptyAttributes = options.stripEmptyAttributes || DEFAULT_EMPTY_ATTRS;
+  const escapeAttributes = /(&|")/g;
+  /** @param {string} match */
+  const escapeAttributesFn = match => (match === '&' ? '&amp;' : '&quot;');
 
   let text = '';
   let depth = -1;
@@ -99,8 +102,18 @@ export function getDiffableHTML(html, options = {}) {
   }
 
   /** @param {Node} node */
+  function getTagName(node) {
+    // Use original tag if available via data-tag-name attribute (use-case for scoped elements)
+    // See packages/scoped-elements for more info
+    if (node instanceof Element) {
+      return node.getAttribute('data-tag-name') || node.localName;
+    }
+    return node.nodeName.toLowerCase();
+  }
+
+  /** @param {Node} node */
   function shouldProcessChildren(node) {
-    const name = node.nodeName.toLowerCase();
+    const name = getTagName(node);
     return (
       !ignoreTags.includes(name) &&
       !ignoreChildren.includes(name) &&
@@ -129,7 +142,7 @@ export function getDiffableHTML(html, options = {}) {
   function getAttributeString(el, { name, value }) {
     if (shouldStripAttribute({ name, value })) return '';
     if (name === 'class') return ` class="${getClassListValueString(el)}"`;
-    return ` ${name}="${value}"`;
+    return ` ${name}="${value.replace(escapeAttributes, escapeAttributesFn)}"`;
   }
 
   /**
@@ -148,7 +161,7 @@ export function getDiffableHTML(html, options = {}) {
             `An object entry to ignoreAttributes should contain a 'tags' and an 'attributes' property.`,
           );
         }
-        return e.tags.includes(el.nodeName.toLowerCase()) && e.attributes.includes(attr.name);
+        return e.tags.includes(getTagName(el)) && e.attributes.includes(attr.name);
       });
     };
   }
@@ -176,13 +189,13 @@ export function getDiffableHTML(html, options = {}) {
 
   /** @param {Element} el */
   function printOpenElement(el) {
-    text += `${getIndentation()}<${el.localName}${getAttributesString(el)}>\n`;
+    text += `${getIndentation()}<${getTagName(el)}${getAttributesString(el)}>\n`;
   }
 
   /** @param {Node} node */
   function onNodeStart(node) {
     // don't print this node if we should ignore it
-    if (node.nodeName === 'DIFF-CONTAINER' || ignoreTags.includes(node.nodeName.toLowerCase())) {
+    if (getTagName(node) === 'diff-container' || ignoreTags.includes(getTagName(node))) {
       return;
     }
 
@@ -204,17 +217,17 @@ export function getDiffableHTML(html, options = {}) {
 
   /** @param {Element} el */
   function printCloseElement(el) {
-    if (el.localName === 'diff-container' || VOID_ELEMENTS.includes(el.localName)) {
+    if (getTagName(el) === 'diff-container' || VOID_ELEMENTS.includes(getTagName(el))) {
       return;
     }
 
-    text += `${getIndentation()}</${el.localName}>\n`;
+    text += `${getIndentation()}</${getTagName(el)}>\n`;
   }
 
   /** @param {Node} node */
   function onNodeEnd(node) {
     // don't print this node if we should ignore it
-    if (ignoreTags.includes(node.nodeName.toLowerCase())) {
+    if (ignoreTags.includes(getTagName(node))) {
       return;
     }