+++ packages/dedupe-mixin/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.3.0](https://github.com/open-wc/open-wc/compare/@open-wc/dedupe-mixin@1.2.18...@open-wc/dedupe-mixin@1.3.0) (2020-08-05)
+
+
+### Features
+
+* **dedupe-mixin:** improve performance by simplifying the code ([#1708](https://github.com/open-wc/open-wc/issues/1708)) ([0ac7fff](https://github.com/open-wc/open-wc/commit/0ac7fffe26fffb5bcd8dcf457b185774efb2d395))
+
+
+
+
+
 ## [1.2.18](https://github.com/open-wc/open-wc/compare/@open-wc/dedupe-mixin@1.2.17...@open-wc/dedupe-mixin@1.2.18) (2020-07-08)