+++ packages/scoped-elements/CHANGELOG.md
@@ -3,6 +3,39 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.2](https://github.com/open-wc/open-wc/compare/@open-wc/scoped-elements@1.2.1...@open-wc/scoped-elements@1.2.2) (2020-08-19)
+
+
+### Bug Fixes
+
+* **scoped-elements:** add host to the mixin type for static props ([88ffd99](https://github.com/open-wc/open-wc/commit/88ffd995efeb94d79ee686f665d18e8ca405e3bc))
+
+
+
+
+
+## [1.2.1](https://github.com/open-wc/open-wc/compare/@open-wc/scoped-elements@1.2.0...@open-wc/scoped-elements@1.2.1) (2020-08-14)
+
+
+### Bug Fixes
+
+* **scoped-elements:** make code ie11 compatible ([08483df](https://github.com/open-wc/open-wc/commit/08483df864276f7d1ac5a72ca0fc1b37ad6367df))
+
+
+
+
+
+# [1.2.0](https://github.com/open-wc/open-wc/compare/@open-wc/scoped-elements@1.1.2...@open-wc/scoped-elements@1.2.0) (2020-08-05)
+
+
+### Features
+
+* **scoped-elements:** add data-tag-name with the original tagName ([d97454e](https://github.com/open-wc/open-wc/commit/d97454ea7e39457b754ee907c6d686aba5d5078b))
+
+
+
+
+
 ## [1.1.2](https://github.com/open-wc/open-wc/compare/@open-wc/scoped-elements@1.1.1...@open-wc/scoped-elements@1.1.2) (2020-07-08)
 
 **Note:** Version bump only for package @open-wc/scoped-elements