+++ packages/semantic-dom-diff/CHANGELOG.md
@@ -3,6 +3,39 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.19.0](https://github.com/open-wc/open-wc/compare/@open-wc/semantic-dom-diff@0.18.0...@open-wc/semantic-dom-diff@0.19.0) (2020-10-03)
+
+
+### Features
+
+* **semantic-dom-diff:** diff options now process scoped elements ([caa1608](https://github.com/open-wc/open-wc/commit/caa16081e38cc4202a8300cf5637657a6ff11a3d))
+
+
+
+
+
+# [0.18.0](https://github.com/open-wc/open-wc/compare/@open-wc/semantic-dom-diff@0.17.10...@open-wc/semantic-dom-diff@0.18.0) (2020-09-11)
+
+
+### Features
+
+* **semantic-dom-diff:** diff Scoped Element tags correctly into Shadow DOM snapshots ([4385957](https://github.com/open-wc/open-wc/commit/43859579bd6131e205130cdf90021fce3fd94b20))
+
+
+
+
+
+## [0.17.10](https://github.com/open-wc/open-wc/compare/@open-wc/semantic-dom-diff@0.17.9...@open-wc/semantic-dom-diff@0.17.10) (2020-08-27)
+
+
+### Bug Fixes
+
+* **semantic-dom-diff:** fix unescaped attributes ([375180d](https://github.com/open-wc/open-wc/commit/375180de30abef00a05ac84bb30f62555f0d7059))
+
+
+
+
+
 ## [0.17.9](https://github.com/open-wc/open-wc/compare/@open-wc/semantic-dom-diff@0.17.8...@open-wc/semantic-dom-diff@0.17.9) (2020-04-26)