+++ packages/testing-helpers/CHANGELOG.md
@@ -3,6 +3,63 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.8.10](https://github.com/open-wc/open-wc/compare/@open-wc/testing-helpers@1.8.9...@open-wc/testing-helpers@1.8.10) (2020-10-02)
+
+
+### Bug Fixes
+
+* **testing-helpers:** replace ts-expect-error with ignore ([f64ed31](https://github.com/open-wc/open-wc/commit/f64ed31c4595019636ff8afc249fb9c01b858522))
+
+
+
+
+
+## [1.8.9](https://github.com/open-wc/open-wc/compare/@open-wc/testing-helpers@1.8.8...@open-wc/testing-helpers@1.8.9) (2020-08-19)
+
+
+### Bug Fixes
+
+* **scoped-elements:** add host to the mixin type for static props ([88ffd99](https://github.com/open-wc/open-wc/commit/88ffd995efeb94d79ee686f665d18e8ca405e3bc))
+
+
+
+
+
+## [1.8.8](https://github.com/open-wc/open-wc/compare/@open-wc/testing-helpers@1.8.7...@open-wc/testing-helpers@1.8.8) (2020-08-16)
+
+**Note:** Version bump only for package @open-wc/testing-helpers
+
+
+
+
+
+## [1.8.7](https://github.com/open-wc/open-wc/compare/@open-wc/testing-helpers@1.8.6...@open-wc/testing-helpers@1.8.7) (2020-08-14)
+
+**Note:** Version bump only for package @open-wc/testing-helpers
+
+
+
+
+
+## [1.8.6](https://github.com/open-wc/open-wc/compare/@open-wc/testing-helpers@1.8.5...@open-wc/testing-helpers@1.8.6) (2020-08-05)
+
+**Note:** Version bump only for package @open-wc/testing-helpers
+
+
+
+
+
+## [1.8.5](https://github.com/open-wc/open-wc/compare/@open-wc/testing-helpers@1.8.4...@open-wc/testing-helpers@1.8.5) (2020-08-04)
+
+
+### Bug Fixes
+
+* **testing-helpers:** change default timeout to 1000ms for waitUntil ([dacf46f](https://github.com/open-wc/open-wc/commit/dacf46faccc9939e26902ef4fbbb102418b81ba2))
+
+
+
+
+
 ## [1.8.4](https://github.com/open-wc/open-wc/compare/@open-wc/testing-helpers@1.8.3...@open-wc/testing-helpers@1.8.4) (2020-07-08)
 
 **Note:** Version bump only for package @open-wc/testing-helpers