+++ packages/testing-karma/CHANGELOG.md
@@ -3,6 +3,62 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.0.8](https://github.com/open-wc/open-wc/compare/@open-wc/testing-karma@4.0.7...@open-wc/testing-karma@4.0.8) (2020-10-03)
+
+**Note:** Version bump only for package @open-wc/testing-karma
+
+
+
+
+
+## [4.0.7](https://github.com/open-wc/open-wc/compare/@open-wc/testing-karma@4.0.6...@open-wc/testing-karma@4.0.7) (2020-10-01)
+
+**Note:** Version bump only for package @open-wc/testing-karma
+
+
+
+
+
+## [4.0.6](https://github.com/open-wc/open-wc/compare/@open-wc/testing-karma@4.0.5...@open-wc/testing-karma@4.0.6) (2020-09-25)
+
+**Note:** Version bump only for package @open-wc/testing-karma
+
+
+
+
+
+## [4.0.5](https://github.com/open-wc/open-wc/compare/@open-wc/testing-karma@4.0.4...@open-wc/testing-karma@4.0.5) (2020-08-27)
+
+**Note:** Version bump only for package @open-wc/testing-karma
+
+
+
+
+
+## [4.0.4](https://github.com/open-wc/open-wc/compare/@open-wc/testing-karma@4.0.3...@open-wc/testing-karma@4.0.4) (2020-08-19)
+
+**Note:** Version bump only for package @open-wc/testing-karma
+
+
+
+
+
+## [4.0.3](https://github.com/open-wc/open-wc/compare/@open-wc/testing-karma@4.0.2...@open-wc/testing-karma@4.0.3) (2020-08-10)
+
+**Note:** Version bump only for package @open-wc/testing-karma
+
+
+
+
+
+## [4.0.2](https://github.com/open-wc/open-wc/compare/@open-wc/testing-karma@4.0.1...@open-wc/testing-karma@4.0.2) (2020-08-04)
+
+**Note:** Version bump only for package @open-wc/testing-karma
+
+
+
+
+
 ## [4.0.1](https://github.com/open-wc/open-wc/compare/@open-wc/testing-karma@4.0.0...@open-wc/testing-karma@4.0.1) (2020-07-29)
 
 **Note:** Version bump only for package @open-wc/testing-karma